+++ src/app/api/menu-items/route.ts
@@ -1,13 +1,17 @@
 import { MenuItem } from "@/app/models/MenuItem";
 import mongoose from "mongoose";
 import { NextRequest, NextResponse } from "next/server";
+import { isAdmin } from "../auth/[...nextauth]/route";
 
 export async function POST(req: NextRequest) {
   try {
     mongoose.connect(process.env.MONGODB_URI!);
-    const data = await req.json();
-    const menuItem = await MenuItem.create(data);
-    return NextResponse.json(menuItem);
+    if (await isAdmin()) {
+      const data = await req.json();
+      const menuItem = await MenuItem.create(data);
+      return NextResponse.json(menuItem);
+    }
+    return NextResponse.json({});
   } catch (err: any) {
     return NextResponse.json(err);
   }
@@ -16,9 +20,12 @@ export async function POST(req: NextRequest) {
 export async function PUT(req: NextRequest) {
   try {
     mongoose.connect(process.env.MONGODB_URI!);
-    const {_id, ...data} = await req.json();
-    const updatedMenuItem = await MenuItem.findByIdAndUpdate({ _id }, data, { new: true });
-    return NextResponse.json(updatedMenuItem);
+    if (await isAdmin()) {
+      const { _id, ...data } = await req.json();
+      const updatedMenuItem = await MenuItem.findByIdAndUpdate({ _id }, data, { new: true });
+      return NextResponse.json(updatedMenuItem);
+    }
+    return NextResponse.json({});
   } catch (err: any) {
     return NextResponse.json(err);
   }
@@ -30,13 +37,16 @@ export async function GET() {
   return NextResponse.json(menuItems);
 }
 
-export async function DELETE(req: NextRequest) { 
+export async function DELETE(req: NextRequest) {
   try {
     mongoose.connect(process.env.MONGODB_URI!);
-    const url = new URL(req.url);
-    const _id = url.searchParams.get('_id');
-    const deleteResult = await MenuItem.deleteOne({ _id });
-    return NextResponse.json(deleteResult);
+    if (await isAdmin()) {
+      const url = new URL(req.url);
+      const _id = url.searchParams.get('_id');
+      const deleteResult = await MenuItem.deleteOne({ _id });
+      return NextResponse.json(deleteResult);
+    }
+    return NextResponse.json(true);
   } catch (err) {
     return NextResponse.json(err);
   }
